requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
blank==0.1.0
category==0.1.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
numpy==1.22.3
Pillow==9.0.1
python-decouple==3.6
pytz==2021.3
rust-category==0.2.0
sqlparse==0.4.2
